refactor: #390 `Place` business logic을 CQRS pattern 적용하여 분리
